+++ src/quacc/recipes/emt/defects.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 from pymatgen.analysis.defects.generators import VacancyGenerator
 
-from quacc import flow, job, subflow
+from quacc import flow, subflow
 from quacc.atoms.defects import make_defects_from_bulk
 from quacc.recipes.emt.core import relax_job, static_job
 
@@ -81,8 +81,7 @@ def bulk_to_defects_flow(
     if "relax_cell" not in defect_relax_kwargs:
         defect_relax_kwargs["relax_cell"] = False
 
-    @job
-    def _make_defects(atoms):
+    def _make_defects(atoms: Atoms) -> list[Atoms]:
         return make_defects_from_bulk(
             atoms,
             defect_gen=defect_gen,
@@ -91,11 +90,13 @@ def bulk_to_defects_flow(
         )
 
     @subflow
-    def _relax_distributed(defects):
+    def _relax_distributed(atoms: Atoms) -> list[OptSchema]:
+        defects = _make_defects(atoms)
         return [relax_job(defect, **defect_relax_kwargs) for defect in defects]
 
     @subflow
-    def _relax_and_static_distributed(defects):
+    def _relax_and_static_distributed(atoms: Atoms) -> list[RunSchema]:
+        defects = _make_defects(atoms)
         return [
             static_job(
                 relax_job(defect, **defect_relax_kwargs)["atoms"],
@@ -104,10 +105,8 @@ def bulk_to_defects_flow(
             for defect in defects
         ]
 
-    defects = _make_defects(atoms)
-
     return (
-        _relax_and_static_distributed(defects)
+        _relax_and_static_distributed(atoms)
         if run_static
-        else _relax_distributed(defects)
+        else _relax_distributed(atoms)
     )