+++ src/quacc/recipes/emt/slabs.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from quacc import flow, job, subflow
+from quacc import flow, subflow
 from quacc.atoms.slabs import make_slabs_from_bulk
 from quacc.recipes.emt.core import relax_job, static_job
 
@@ -59,16 +59,17 @@ def bulk_to_slabs_flow(
     if "relax_cell" not in slab_relax_kwargs:
         slab_relax_kwargs["relax_cell"] = False
 
-    @job
-    def _make_slabs(atoms):
+    def _make_slabs(atoms: Atoms) -> list[Atoms]:
         return make_slabs_from_bulk(atoms, **make_slabs_kwargs)
 
     @subflow
-    def _relax_distributed(slabs):
+    def _relax_distributed(atoms: Atoms) -> list[OptSchema]:
+        slabs = _make_slabs(atoms)
         return [relax_job(slab, **slab_relax_kwargs) for slab in slabs]
 
     @subflow
-    def _relax_and_static_distributed(slabs):
+    def _relax_and_static_distributed(atoms: Atoms) -> list[OptSchema]:
+        slabs = _make_slabs(atoms)
         return [
             static_job(
                 relax_job(slab, **slab_relax_kwargs)["atoms"],
@@ -77,10 +78,8 @@ def bulk_to_slabs_flow(
             for slab in slabs
         ]
 
-    slabs = _make_slabs(atoms)
-
     return (
-        _relax_and_static_distributed(slabs)
+        _relax_and_static_distributed(atoms)
         if run_static
-        else _relax_distributed(slabs)
+        else _relax_distributed(atoms)
     )